git-svn-id: http://svn.osafoundation.org/chandler/trunk@8981 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/chandlerdb/chandlerdb/item/item.h
@@ -16,9 +16,9 @@
 #define _ITEM_H
 
 
-#define Item_HEAD          \
-    unsigned long status;  \
-    unsigned long version; \
+#define Item_HEAD                   \
+    unsigned long status;           \
+    unsigned long long version;     \
     PyObject *name;
